Move some functions to separate utilities module.
